Osint Automator/requirements.txt
requests
dnspython
colorama
pyfiglet
